:sparkles: refactor: changing struct
