+++ src/main/java/br/ifsp/husaocarlos/domain/entities/Action.java
@@ -1,0 +1,4 @@
+package br.ifsp.husaocarlos.domain.entities;
+
+public class Action {
+}
