+++ src/main/java/br/ifsp/husaocarlos/domain/entities/Patient.java
@@ -1,0 +1,4 @@
+package br.ifsp.husaocarlos.domain.entities;
+
+public class Patient{
+}
