--- src/main/java/br/ifsp/husaocarlos/model/Action.java
@@ -1,4 +1,0 @@
-package br.ifsp.husaocarlos.model;
-
-public class Action {
-}
