--- src/main/java/br/ifsp/husaocarlos/model/Patient.java
@@ -1,4 +1,0 @@
-package br.ifsp.husaocarlos.model;
-
-public class Patient{
-}
